AB-006 - Ablility to clear current value in field Bug: Error while updating with same value

+++ src/main/groovy/com/khabaznia/bots/core/service/BotMessagesService.groovy
@@ -4,6 +4,7 @@ import com.khabaznia.bots.core.meta.Emoji
 import com.khabaznia.bots.core.meta.container.DefaultRequestContainer
 import com.khabaznia.bots.core.meta.keyboard.impl.InlineKeyboard
 import com.khabaznia.bots.core.meta.request.BaseRequest
+import com.khabaznia.bots.core.meta.request.impl.SendMessage
 import com.khabaznia.bots.core.trait.BaseRequests
 import com.khabaznia.bots.core.trait.Configurable
 import com.khabaznia.bots.core.trait.Loggable
@@ -11,10 +12,10 @@ import groovy.util.logging.Slf4j
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Component
 
-import static com.khabaznia.bots.core.controller.Constants.EDIT_FIELD_CONTROLLER.EDIT_BOOLEAN_FIELD
-import static com.khabaznia.bots.core.controller.Constants.EDIT_FIELD_CONTROLLER.EDIT_LOCALIZED_FIELD_MENU
+import static com.khabaznia.bots.core.controller.Constants.EDIT_FIELD_CONTROLLER.*
 import static com.khabaznia.bots.core.controller.Constants.LANG_CONTROLLER.LANG_EMOJI
 import static com.khabaznia.bots.core.flow.service.EditFlowService.enterMessage
+import static com.khabaznia.bots.core.flow.service.EditFlowService.isValueClearingEnabled
 import static com.khabaznia.bots.core.routing.Constants.AVAILABLE_LOCALES
 import static com.khabaznia.bots.core.util.SessionUtil.currentUser
 import static org.apache.groovy.parser.antlr4.util.StringUtils.isEmpty
@@ -34,9 +35,13 @@ class BotMessagesService implements BaseRequests, Loggable, Configurable {
     }
 
     void editFlowEnterMessage(String text, Map<String, String> binding) {
-        requests << sendMessage.text(text ?: enterMessage)
+        def message = sendMessage.text(text ?: enterMessage ?: getDefaultEditFlowEnterMessage())
                 .binding(binding)
-                .delete()
+                .delete() as SendMessage
+        if (isValueClearingEnabled() && !isEmpty(currentUser.editFlow.oldValue?.strip())) {
+            message.keyboard(inlineKeyboard.button('button.edit.flow.clear.value', EDIT_FIELD_CLEAR_VALUE))
+        }
+        requests << message
     }
 
     void editFlowCurrentValueMessage(String currentValue, boolean isBooleanValue = false) {
@@ -46,10 +51,11 @@ class BotMessagesService implements BaseRequests, Loggable, Configurable {
     }
 
     void updateEditFlowCurrentValueMessage(String currentValue, boolean isBooleanValue = false) {
-        requests << editMessage.text(getEditFlowCurrentValueText(currentValue))
-                .binding([value: getEditFlowMappedBooleanValue(isBooleanValue, currentValue)])
-                .label(currentUser.code.concat(EDIT_FLOW_CURRENT_VALUE_MESSAGE))
-                .delete()
+        if (currentUser.editFlow.oldValue != currentValue)
+            requests << editMessage.text(getEditFlowCurrentValueText(currentValue))
+                    .binding([value: getEditFlowMappedBooleanValue(isBooleanValue, currentValue)])
+                    .label(currentUser.code.concat(EDIT_FLOW_CURRENT_VALUE_MESSAGE))
+                    .delete()
     }
 
     void editFlowChooseLangMessage() {
@@ -63,7 +69,6 @@ class BotMessagesService implements BaseRequests, Loggable, Configurable {
                 .binding([lang: LANG_EMOJI[lang]])
                 .label(currentUser.code.concat(EDIT_FLOW_UPDATE_MESSAGE))
                 .delete()
-
     }
 
     void deleteEditFlowChooseLangMessage() {
@@ -78,8 +83,8 @@ class BotMessagesService implements BaseRequests, Loggable, Configurable {
                 .delete()
     }
 
-    void editFlowSuccessMessage(String text) {
-        requests << sendMessage.text(text ?: 'text.edit.flow.success.message')
+    void editFlowSuccessMessage(String text, boolean clear = false) {
+        requests << sendMessage.text(text ?: (clear ? 'text.edit.flow.cleared.message' : 'text.edit.flow.success.message'))
                 .delete()
     }
 
@@ -101,5 +106,9 @@ class BotMessagesService implements BaseRequests, Loggable, Configurable {
 
     private static String getEditFlowMappedBooleanValue(boolean isBooleanValue, String currentValue) {
         isBooleanValue ? BOOLEAN_VALUES_MAPPING.get(currentValue) : currentValue
+    }
+
+    private static String getDefaultEditFlowEnterMessage() {
+        isValueClearingEnabled() ? 'text.edit.flow.enter.new.value.or.clear' : 'text.edit.flow.enter.new.value'
     }
 }
